chore: move lib/gen to lib/src/gen
