+++ .gitignore
@@ -26,7 +26,7 @@ migrate_working_dir/
 .pub/
 build/
 coverage/
-/packages/*/lib/gen/
+/packages/*/lib/src/gen/
 pubspec_overrides.yaml
 
 # Generated files
