+++ packages/analysis_options/lib/analysis_options.yaml
@@ -11,7 +11,7 @@ include: package:very_good_analysis/analysis_options.yaml
 
 analyzer:
   exclude:
-    - "lib/gen/**"
+    - "lib/src/gen/**"
     - "lib/src/l10n/app_localizations*.dart"
     - "lib/generated_plugin_registrant.dart"
     - "test/.test_coverage.dart"
