+++ packages/app/lib/src/features/auth/presentation/auth/login_page.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../../../../gen/assets.gen.dart';
 import '../../../../app/router.gr.dart';
+import '../../../../gen/assets.gen.dart';
 import '../../application/auth_service.dart';
 
 // TODO(lishaduck): Rename to `LogInPage`.
